+++ src/platform.h
@@ -11,7 +11,7 @@ typedef struct t_ts_file_content
 {
 	size_t content_length;
 	void *content;
-	int file_error;
+	uint16_t file_error;
 } ts_file_content;
 
 typedef enum t_ts_file_open_error
